personalised version of XioScript
I’ve just learned what a pull request on github is. I’ve "forked" the original script, it created a respository looking like the original one but titled "TeaMike/XioScript". I've edited it (enabled developer mode and @excluded some functions.

This is a test with the intention to create a pull request to see how this works.

To commit my changes i am presented with two choices:

 + Commit directly to the master branch
 +  Create a new branch for this commit and start a pull request. Learn more about pull requests.