Version 12.0.7
Fixed a bug where the title "XS 12 Maintenance Log" was shown twice or
more if the Maintenance button was pressed again without reloading.
Fixed a bug where the top manager employees on profile qualification
maximum used the values for target.
Fixed a bug where the available number of goods in the script did not
match the virtonomics values if a limit was set.
Fixed a bug where the new suppliers counter did not count the number of
new suppliers but the number of deleted suppliers.
Update for the new supply interface