Version 12.0.15
The CTIE and Profit Tax price function now sets the price to prime cost
instead of 30*IP if the price is higher than 30*IP.
Fixed a bug where the warehouse supply market function would not be
deleted a contract if a new price in that contract was set.
Removed Copy and Paste functions
Added click and drag selection
The 'continue research' Research function now gives off a message if a
new technology level could not be set due to the lab size.
Added step2 for the 'continue research' Research function
Added 'Market PQR' Production Price function
Added '1 year training' Training function
Fixed a bug where filtering on subdivision type with the
virtonomics-built functions on the unit list caused removals of options.